Replace `HashedMediaStorage` with `UniqueStorage` and `UniqueMixin`.
This new storage class *always* uses just a content hash as the filename
and *always* stores files in a special directory for deduplicated files,
but retains the original path within the deduplicated directory, so
we can still group files for a single field together and avoid mixing
original and deduplicated files in a single directory.

The default deduplicated directory is `dd`, and this can be overridden
with the `IXC_WHITENOISE_DEDUPE_PATH_PREFIX` setting.

We also have a `deduplicate_unique_storage` management command, which
finds ALL file fields that use `UniqueStorage` and re-saves them to
deduplicate. The original files are not removed.

Every time a unique file is saved, a `UniqueFile` object is created
which stores the original and unique filename. You can use this to
delete the originals if they are no longer required, or force downloads
to use the original name via the `content-disposition` header.

File extensions are always forced to lowercase, and common 4 character
extensions (jpeg, yaml) are replaced with their abbreviated 3 character
versions, to further reduce the probability of duplication occurring.

You can add to the list of extension replacements with the
`IXC_WHITENOISE_DEDUPE_EXTENSIONS` setting, which should be a dict with
4 character extensions as the key (lowercase, with leading period) and
3 character versions as the value.