Fix Django 1.6 compatibility.
`get_models()` in Django 1.6 does not support the same shortcut that newer versions of Django do, where `app_label.model_name` can be passed as a single argument.

Instead of a try/except when importing the different versions of `get_model()` and fixing the args to `get_model()` throughout, just import the actual model when we need it instead of at module level to solve circular import issues.