Fix hashing failure when a decoded file is parsed
As documented in the comment, this is a rare occurrence, since the
content provided from hashing is almost always a binary file or a
bytestream.

However, when uploaded files are attached to an e-mail to be sent,
Django's mail module decodes the file into a StringIO if the mimetype
starts with "text/" (instead of a BytesIO for other mimetypes), which
ends up being passed to storage for saving when the e-mail is sent using
post_office.

In this scenario, it's guaranteed to have been decoded using UTF-8, so
it's safe to encode it as such. Since this is the first time we've
observed such an issue in years, it's fairly safe to hope that in the
unlikely case of some new scenario triggering this, the encoding will
still be UTF-8.